requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fpdf==1.7.2
idna==3.4
-e git+ssh://git@github.com/MarioChvx/manger.git@7c7099963d30baecb8f71cddd7d908162273a2ae#egg=manger
Pillow==9.4.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
